move functions to streams.py
